requirements.txt
fastapi[all]
uvicorn[standard]
python-dotenv
pymongo[srv]
langchain-google-genai
langgraph
requests